procedure to encapsulate average student marks

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -8,7 +8,7 @@ CREATE PROCEDURE ComputeAverageScoreForUser (
   BEGIN
     DECLARE total FLOAT;
     DECLARE student_avg FLOAT;
-    DECLARE num_scores INT;
+    DECLARE num_scores FLOAT;
 
     SELECT SUM(score) INTO total
     FROM corrections
